requirements.txt
-i https://pypi.org/simple
pillow==7.1.2
